% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_transformation.R
\name{clr_transformation}
\alias{clr_transformation}
\title{CLR Transformation}
\usage{
clr_transformation(matrix)
}
\arguments{
\item{matrix}{Input count matrix with samples on rows and species in columns}
}
\value{
matrix. A clr transformed matrix class with samples on rows and species
 in columns
}
\description{
Util function to transform raw counts using centered log ratio transform
approximation (the same approach as the microbiome package used in the
experiments)
}
\examples{
# Original data
print(example_matrix)
em_clr <- clr_transformation(example_matrix)
# Transformed data
print(em_clr)
}
