% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pseq}
\alias{example_pseq}
\title{Example phyloseq for use functions}
\format{
A phyloseq object: otu_table (1657x22),tax_table (22x7) and
 sam_data (1657x25)
}
\usage{
data(example_pseq)
}
\description{
Example phyloseq for use functions
}
\keyword{data}
