% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Get Clusters}
\usage{
get_clusters(object)
}
\arguments{
\item{object}{(phyloseq, matrix or data frame). In the case of the latter two,
the samples in the rows and the species in the columns}
}
\value{
object (phyloseq, matrix or data frame). Same input class object with
with the probabilities and the cluster membership label attached
}
\description{
Main function: from an input object (which can be a matrix, a data frame or
a phyloseq), this function returned the same object with the probabilities
and the cluster membership label attached
}
\examples{
# Original sample data (phyloseq case)
print(phyloseq::sample_data(example_pseq))
new_pseq <- get_clusters(example_pseq)
# New sample data (phyloseq case)
print(phyloseq::sample_data(new_pseq))
 # Original sample data (matrix case)
print(example_matrix)
new_matrix <- get_clusters(example_matrix)
# New sample data (matrix case)
print(new_matrix)
}
