% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VIBES.R
\name{get_VIBES}
\alias{get_VIBES}
\title{Get VIBES}
\usage{
get_VIBES(object, column = NULL)
}
\arguments{
\item{object}{(phyloseq, matrix or data frame). In the case of the latter two,
the samples in the rows and the species in the columns}

\item{column}{(only for phyloseq) Input column number occupied by the taxonomic rank Species. Default is Null and take last column}
}
\value{
object (phyloseq, matrix or data frame). Same input class object with
with the probabilities and the cluster membership label attached
}
\description{
This function uses an elastic net model (trained with the knowledge extracted
from previous cluster analyses) to predict the membership of the input samples
to the VIBES clusters. Based on the 20 species raw counts, it assigns the samples
to the VIBES clusters.
}
\examples{
# Original sample data (phyloseq case)
print(phyloseq::sample_data(PRJNA208535))
new_pseq <- get_VIBES(PRJNA208535)
# New sample data (phyloseq case)
print(phyloseq::sample_data(new_pseq))
}
