% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{model_predict}
\alias{model_predict}
\title{Model Predict}
\usage{
model_predict(newdata)
}
\arguments{
\item{newdata}{A clr transformed matrix class with samples on rows and
species on columns}
}
\value{
data frame. A data frame class object with probabilities of belonging to each cluster
as well as the label of the cluster with the highest probability of
membership
}
\description{
Util function to predict sample membership of each cluster class
}
\examples{
# Data from which you want to predict cluster labels
require(VIBES)
em_clr <- VIBES:::clr_transformation(example_matrix)
pred <- VIBES:::model_predict(example_matrix)
# Probabilities and label for each sample
print(pred)
}
