% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA208535}
\alias{PRJNA208535}
\title{Example phyloseq for use functions}
\format{
A phyloseq object generated by our methodology, corresponding to the ENA data with identifier PRJNA208535.
}
\usage{
data(PRJNA208535)
}
\description{
Example phyloseq for use functions
}
\keyword{data}
