% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA208535_matrix}
\alias{PRJNA208535_matrix}
\title{Example matrix for use functions}
\format{
A 1657x22 matrix with species in columns and samples in rows
}
\usage{
data(PRJNA208535_matrix)
}
\description{
Example matrix for use functions
}
\keyword{data}
