% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA302078}
\alias{PRJNA302078}
\title{Example phyloseq for use functions}
\format{
A phyloseq object generated by our methodology, corresponding to the ENA data with identifier PRJNA302078.
}
\usage{
data(PRJNA302078)
}
\description{
Example phyloseq for use functions
}
\keyword{data}
