% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA797778}
\alias{PRJNA797778}
\title{Example phyloseq for use functions}
\format{
A phyloseq object generated by our methodology, corresponding to the ENA data with identifier PRJNA797778.
}
\usage{
data(PRJNA797778)
}
\description{
Example phyloseq for use functions
}
\keyword{data}
