% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRA022855}
\alias{SRA022855}
\title{Example phyloseq for use functions}
\format{
A phyloseq object generated by our methodology, corresponding to the ENA data with identifier SRA022855.
}
\usage{
data(SRA022855)
}
\description{
Example phyloseq for use functions
}
\keyword{data}
