% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRA051298}
\alias{SRA051298}
\title{Example phyloseq for use functions}
\format{
A phyloseq object generated by our methodology, corresponding to the ENA data with identifier SRA051298.
}
\usage{
data(SRA051298)
}
\description{
Example phyloseq for use functions
}
\keyword{data}
