% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIBES-package.R
\docType{package}
\name{VIBES-package}
\alias{VIBES}
\alias{VIBES-package}
\title{VIBES: VagInal Bacterial subtyping using machine learning for Enhanced classification of bacterial vaginosiS}
\description{
Model based on the glmnet algorithm to cluster the vaginal microbiome profile from only 20 species.
}
\author{
\strong{Maintainer}: Diego Fernández-Edreira \email{diego.fedreira@udc.es} [copyright holder]

Other contributors:
\itemize{
  \item Josel Liñares-Blanco \email{j.linares@udc.es} [contributor]
  \item Carlos Fernandez-Lozano \email{carlos.fernandez@udc.es} [contributor]
}

}
\keyword{internal}
