% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Get Clusters}
\usage{
get_clusters(object, column = NULL)
}
\arguments{
\item{object}{(phyloseq, matrix or data frame). In the case of the latter two,
the samples in the rows and the species in the columns}

\item{column}{(only for phyloseq) Input column number occupied by the taxonomic rank Species. Default is Null and take last column}
}
\value{
object (phyloseq, matrix or data frame). Same input class object with
with the probabilities and the cluster membership label attached
}
\description{
Main function: from an input object (which can be a matrix, a data frame or
a phyloseq), this function returned the same object with the probabilities
and the cluster membership label attached
}
\examples{
# Original sample data (phyloseq case)
print(phyloseq::sample_data(PRJNA208535))
new_pseq <- get_clusters(PRJNA208535)
# New sample data (phyloseq case)
print(phyloseq::sample_data(new_pseq))
}
