% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data (from Phyloseq object)}
\usage{
get_data(pseq)
}
\arguments{
\item{pseq}{Input phyloseq object}
}
\value{
matrix. A matrix class with samples on rows and species in columns
}
\description{
Util function to extract a matrix class with samples on rows and species on
columns from a phyloseq object
}
\examples{
# Original ASV table
require(VMC)
m <- VMC:::get_data(example_pseq)
# Output matrix with taxas names instead ASVs names
print(m)
}
